fix cardOrder when modifying block ids
